Give the DocumentSplitsView absolute control
Till now, the DocumentController needed to know about the underlying
OakDocumentView and its OakTextView to function. This made doing splits
impossible because the DocumentController would get confused. Now, the
DocumentController gets the DocumentView and TextView from the
SplitsView and does not store this information so the SplitsView is
free to manage the documents in whatever way it likes.