Set default position of new split to be the center
When a new split view is created, the new split will be set to half of
it’s available size. Unfortunately doing this requires some trickery,
thus the center function is called until it works. There should
probably be a way to figure out when to call the centering function to
guarantee that it works, but this just gets it done.