Enable arbitrary splitting
When splitting the current active document, if the split is in the same
direction as the parent split view, then the new split is added as just
another split. If it is a different orientation, then the current
document is enclosed in a new split view with the correct orientation,
then the new document is added to this new split view.

When removing splits, the goal is to keep at least 2 documents per
NSSplitView. If removing a document will cause an NSSplitView to only
have 1 document, then the split view is removed and the document is
moved into the parent split view. Also, if the remaining documents in a
split view is in the same orientation as its parent, then the extra
split view is redundant, so the document views are removed from it and
added to the parent and the redundant split view is removed.