Replace the disposable document in all splits
When a document is opened over a disposable document, the disposable
document is thrown away. The new document is opened and all is ignored
if there is only one editor, but with multiple editors open, mark the
document for closing and open the new document in all of the split
views.