Add option to quietly set a document in a TextView
The find windows simply (though very indirectly) sets the document to
the current document, this causes the selection to be moved into view,
this is nice. When the active split view is changed, the document is
set to the current document to update the current document settings and
properties and whatnot, this is also nice. The problem is that simply
changing the active split view should not change the scroll position in
the document, this is not nice. The `quietly` parameter just controls
updating the scroll position when the current document is set to itself.