Don't activate the file browser on double click
Double (multi) clicking an item in the file browser should not make it
active. This caused problems with the split views implementation
because the text view would loose first responder status, then gain it
back before the new document was open. When gaining first responder,
the document controller tries to update the active document, but the
document is not yet open so it falls back to the old document. Once the
new document is opened the tab bar becomes out of sync.